requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
redis==5.0.1
Werkzeug==3.1.6
WTForms==3.2.1
python-dotenv==1.0.0
boto3==1.34.0
botocore==1.34.162
gunicorn==22.0.0
PyJWT==2.12.0
resend==2.5.1
pytest==8.3.2
black==26.3.1
ruff==0.6.4
